Cleanup OIDC Flow
