+++ src/app/components/IdentityProviderWizard/Wizards/Providers/Salesforce/OIDC/Steps/2.tsx
@@ -1,5 +1,6 @@
 import React from "react";
 import SalesforceOidcStep0Image from "@app/images/salesforce/OIDC/salesforce_oidc_0.png";
+import SalesforceOidcStep1Image from "@app/images/salesforce/OIDC/salesforce_oidc_1.png";
 import {
   InstructionProps,
   Step,
@@ -16,10 +17,8 @@ export const SalesforceStepTwo: React.FC<Props> = ({ loginRedirectURL }) => {
     {
       text: (
         <div>
-          Under the <b>API (Enable OAuth Settings)</b> section, check the <b>Enable OAuth Settings</b> checkbox,{" "}
-          uncheck the <b>Require Proof Key for Code Exchange (PKCE) Extension for Supported Authorization Flows</b> checkbox,{" "}
-          add the <b>Access unique user identifiers (openid)</b> scope, and paste the URL from below into the{" "}
-          <b>Callback URL</b> field.
+          Under the <b>API (Enable OAuth Settings)</b> section, check the <b>Enable OAuth Settings</b>{" "}
+          checkbox and paste the URL from below into the <b>Callback URL</b> field.
         </div>
       ),
       component: <StepImage src={SalesforceOidcStep0Image} alt="Step 2.1" />,
@@ -31,6 +30,17 @@ export const SalesforceStepTwo: React.FC<Props> = ({ loginRedirectURL }) => {
           initialValue={loginRedirectURL}
         />
       ),
+    },
+    {
+      text: (
+        <div>
+          Select the <b>Access the identity URL service (id, profile, email, address, phone)</b> and{" "}
+          <b>Access unique user identifiers (openid)</b> OAuth Scopes, then click the <b>Add</b> button.{" "}
+          <br/><br/>
+          Uncheck the <b>Require Proof Key for Code Exchange (PKCE) Extension for Supported Authorization Flows</b> checkbox.
+        </div>
+      ),
+      component: <StepImage src={SalesforceOidcStep1Image} alt="Step 2.1" />,
     },
     {
       component: (
