+++ src/app/components/IdentityProviderWizard/Wizards/Providers/Salesforce/OIDC/Steps/index.ts
@@ -1,3 +1,2 @@
 export * from "./2";
 export * from "./3";
-export * from "./4";
