+++ src/app/configurations/identity-providers.ts
@@ -14,7 +14,7 @@ import samlLogo from "@app/images/provider-logos/saml_logo.svg";
 import vmwareLogo from "@app/images/provider-logos/vmware_logo.svg";
 import ldapLogo from "@app/images/provider-logos/ldap_logo.svg";
 import duoLogo from "@app/images/duo/duo.svg";
-import salesforceLogo from "@app/images/salesforce/salesforce-logo.svg";
+import salesforceLogo from "@app/images/salesforce/salesforce-logo.png";
 import { Protocols } from ".";
 
 export enum Providers {
